Added pipeline to Location Resource.  Removed Pipeline validation
Pipeline validation was causing errors due to external bug.

ModelResources with ForeignKeys to another model don't work with
validation = CleanedDataFormValidation  On creation, it errors with:
"Select a valid choice. That choice is not one of the available
choices." This is because the ModelResource accepts a URI, but does not
convert it to a primary key (in our case, UUID) before passing it to
Django.
See https://github.com/toastdriven/django-tastypie/issues/152 for
details