AIP store aware of AIP Pointer file
refs #5159

When storing an AIP, storage service now also grabs the pointer file
associated with it and stores it locally, updating it for the new AIP
location. A new Location purpose of STORAGE_SERVICE_INTERNAL created to store
the pointer.xml files.  Packages store where the pointer file is located.

lxml added as a dependency for parsing XML files.