Change locations URLs
Change the location delete and detail URLS to not be inside space/.
Locations have their own UUID, and don't need the space's UUID to be
identified, and it only made the URLs more complicated.

Location create is still inside space/<uuid of space> because we need to
know what space a location goes in during creation.  Location edit also must
take a space UUID as a parameter, since it uses the same view as create.