Allow backlog transfers to be stored & deleted
refs #6131

Add support for storing a transfer in a backlog Location when a Package is
created.  Add checks to ensure packages only go to the correct Location's
purpose (ie AIPs go to AIP storage, transfer to transfer backlog). Refactor
store AIP code to share some in common with backlog transfer.  Expect AIP
storage origin_path's to be relative to the Location. Allow transfers to be
deleted from backlog