Update pointer file for entry creation
refs #5425

Update structMap and fileGrp, for when a package is split or unsplit.  Add
event and SS agent when the file is split.

Add helper function for writing PREMIS events and agents.  Update structMap
and fileGrp of pointer file with information on chunks.

Add version.py for tracking storage service version and providing string
display of it.