Update pointer file with LOCKSS URLs
refs #5425

Query the state_iri for this package and parse it for the server states. If
all are in agreement, add those URLs to the pointer file for each LOCKSS
chunk.  Works for both split and unsplit packages.

Factor out generation of download URL, as that is the ID of lom:content
elements in the SWORD statement.