Download URL is absolute
refs #5425

Require the externally available URL to be configured for LOM Spaces, as
programatically determining the IP is unreliable.  Download URLs provided
to Lockss-o-matic prepend this to the relative path.

Service Document IRI converted to URLField to ensure it is a valid URL.
Collection IRI left as CharField, as it is read from the SD, and any
validation error would probably get buried.