Update package status will re-post to LOM
refs #5425

Updating package status will re-submit the atom entry to LOM, if the
first entry failed. post_move_from_storage_service updated to be
idempotent, including all the functions it calls.

Refactored split_package to include all modifications to the pointer
file, and create_resource to only create the atom entry.  split_package
will return the output files if a split has already been performed.

Generate checksum moved to utils.  Parsed pointer file for current
Package made an attribute of Lockssomatic, populated on demand, so it
is easier for functions to reuse it.