Associate deposits with Packages, collections with Locations
A deposit is more closely tied to a Package conceptually - a
file or collection of files to be moved around between Locations.
Locations are more associated with a purpose, and a logical grouping of
Packages.

Update resource endpoints to associate SWORD deposits with Packages,
and collections with Locations.  Move deposit related endpoints to
PackageResource, and pass a Package to the associated SWORD function.
Update DownloadTask & DownloadTaskFile to associate with Packages.
Update DownloadTask status function to return constants.