Call approve reingest endpoint in dashboard
refs #6488

Copy the reingest files to a staging directory on the pipeline, and call a
SIP reingest endpoint to notify Archivematica.  This allows the pipeline to
do some pre-processing of the SIP, such as clearing the database of old
information.

Enable PATCH for pipelines, specificially so that 'reingest' can be reset by
the pipeline.