Package: Split get_local_path to get/fetch_local_path
get_local_path does not imply it may do a copy. Split it into two
functions, get_local_path which returns a local path only if it already
exists (or None), and fetch_local_path with the old behaviour of
copying the package locally if it isn't found. Update references to use
fetch_local_path.