Duracloud: Browse works without leading /
refs #7133

Browse resources should only join with the Space/Location path if it does
not contain the Space/Location path already.  Some calls to browse contain
the full path, and others relative paths.  This should work for Spaces with
or without a leading / in the path.

Update directory browser JS in SS to handle paths that don't begin with / as
well.