Duracloud: set chunk size to 1GB
Workaround for a bug in the DuraCloud REST API that kicks in when the size
of the file being transferred is larger than 2^31 bytes (2.14 GB).
After the bug is fixed, it should be OK to restore to 5GB.

Chunking to keep files in the 1GB range is a good workaround for now, but should not be required for long.