Replace -a flag in rsync to avoid CIFS errors ref. #8124
The purpose of using rsync with -a was to preserve file
timestamps when copying files from/to the storage service.
rsync -a attemps to preserve timestamps and permissions of copied
files and directories. However, on a CIFS mounted volume it
seems not possible to set timestamps and permissions of
copied directories, producing a non-zero return code.
Replace -a with -t and -O. -t will preserve modification
times. -O will avoid to preserve modification times in
copied directories.