Space: don't log exc_info when falling back to browse local
The traceback isn't generally useful here, since the exception being
caught is very specific already. The exception is also misleading;
we've encountered a number of users having problems who believed that
because an exception was printed here, it was the actual cause of their
problems when it was actually unrelated.