remove sword2 and pyOpenSSL from requirements
sword2 cannot be packaged on launchpad, unless it is vendored.
pyOpenSSL has a setup.py file that tries to download dependencies
from the internet, it is not possible to change that behaviour
without forking the library.

This commit moves installation of sword2 and pyOpenSSL to postinst.

testing requirements changes for building packages