Package: disable smart strings for xpath() calls
lxml's xpath() method returns "smart strings" if the XPath expression
returns a string, for example if the expression terminates in text().
These smart strings are extended from the core types by including
getparent() methods which know their originating element:
http://lxml.de/xpathxslt.html

This has a couple of disadvantages:
1) Memory performance is worse, since it means that the tree is being
   unnecessarily kept in memory even if it's not being used.
2) When using PyPy and lxml before 3.5.0, PyPy will crash with a
   stack overflow:
   http://permalink.gmane.org/gmane.comp.python.lxml.devel/7572

Since we're not using the extra behaviour of these strings, it's better
for us to disable it and just return a raw string.