Combine user edit and password change forms
They are now controlled by a single button - this is to prevent users
from accidentally not saving their user data due to not realising saves
must be done separately.

The two forms remain separate (as they are based on Django forms) but
the view will treat them atomically - data will be saved on both forms
if and only if they are both correctly filled in. (Note that in the
case of the password form, correctly filled in can mean empty - the user
may choose not to change their password).

Tests also added.