Package: Fix storing uncompressed reingested AIPs
When storing an uncompressed AIP that has been reingested, delete the old
preservation derivatives from both the AIP being stored and the
destination.  Otherwise, the extra files cause it to fail bagit validation.
This is not a problem with packaged AIPs because the AIP is replaced as a
single unit.

refs redmine 10080