Re-ingest: don't delete pointer file, refs #11356
In #203, a change was introduced in finish_reingest() to stop deleting the
pointer file if the AIP was uncompressed (see #11205 for more details).
The change was made so the file was instead deleted when the AIP was
compressed which is causing a different issue as that file is used later in
the same function so this commit reverts that part of #203.