Getting checksum for Arkivum from pointer file
Issue: attempting to calculate the checksum of a compressed AIP on disk
after moving it from SS to an Arkivum space was causing an error because
AIP was obviously no longer at that location. Bug appears to have arisen
only after the implementation of the move command was altered to default to
Python/OS move over rsync. Refs RM issue 11138.