Delete create_testuser signal handler
This commit removes the signal handler that was responsible for creating the
`test` superuser. It is still possible to create a superuser from the
command-line interface.

The `base` fixture file is updated so it includes a new superuser. Previously,
the tests were relying on user created by the `create_testuser` signal handler.

This relates to #286. The issue will not be closed until we confirm that the
distributions provision the superuser on demand, e.g. packages or Ansible.