move db dependencies out of base requirements
fixes #222

It should not be necessary to install the database pip packages
such as mysqlclient and psycopg2 in a test environment.

This commit duplicates some deps, in local.txt and production.txt
but it allows test environments to exclude these packages.

This helps reduce the size of the test environment and the time it
takes to run tests, since some of these pip packages require
operating system packages to be installed as well.