Ignore errors in shutil.rmtree
Using `shutil.rmtree` was problematic in NFS drives (see "silly rename")
because we're trying to deleted files that are still being used, i.e. Django
doesn't close the file until the response is fully returned to the browser.

The long-term solution is used to defer the cleanup.

This fixes #269. Read the issue for more details.