Allow SIP to be created from encrypted transfer
The _encr_path2key_fingerprint function was hanging when attempting to
query the db for packages whose current_path ended with the provided
encrypted path. Since the encrypted path can be absolute, the correct
behaviour is to find the package whose current path is contained within
the provided encrypted path. Fixes #301.