Use SQL LIKE to match encrypted path to package
The _encr_path2key_fingerprint function now uses raw SQL queries to find
the package whose current path is a substring of a given encrypted path.
Because the SQL concatenation operator is required in order to select
rows where a column value is a substring of a query parameter, and because
SQLite and MySQL have different concatenation operators some complexity
(Django db engine introspection) is added here.