Fix replication: create replica ptr file if master has one
Fixes ``Package.create_replica_pointer_file`` so that replicas are given
pointer files if their masters have them. Previous behaviour was to check if the
replica was a file (i.e., compressed) but that is incorrect because at the time
of the check the replica package is always a directory.