Update Dataverse Storage Service features
This commit updates various features of the Storage Service related
to browsing and transferring Dataverse datasets inside Archivematica.
The most substantial changes are around transfer browsing where the
toggle will now show the contents of a Dataset when a user looks
into a dataset folder.

Creating a location will no longer cause a Django rendering error
when displaying the 'beta feature' string.

Query flexibility has been improved to include subtree. This
provides users with a greater flexibility to restrict datasets
shown in the transfer browser.

The commit enables the unzip of Dataverse Bundles. Once a Bundle has
been downloaded we unzip it in the appropriate location so that this
work does not have to be completed by Archivematica itself.
Archivematica will then manage other types of package as configured
by the user.

There are various code formatting changes in-line with PEP8.

* Resolves #352
* Resolves #355
* Resolves #380
* Resolves archivematica/issues#108
* Resolves archivematica/issues#113
* Resolves artefactual/archivematica#1085
* Resolves artefactual/archivematica#1121