Hide DSpace REST access protocol
The DSpace REST-type space is not ready to be released. This commit
prevents it from being displayed in the "Create Space" GUI by
introducing ``Space.BETA_PROTOCOLS`` and ``SpaceForm.filter_beta_protocols``.
This approach is better than the approach of commenting out the ``DSPACE_REST``
because that latter approach would require undoing database migrations and
planning to redo them at a later date.