DSpace REST form: mark HTML as safe
Adds ``django.utils.safestring.mark_safe`` so that ``DSpaceRESTForm.as_p``
returns a ``django.utils.safestring.SafeText`` instance instead of a string.
Previously (before this was implemented), if the form had to be
re-rendered because it was invalid or if the user were trying to edit an
existing DSpace REST space, then the HTML would be escaped and would be
displayed to the user as unusable text.

Connected to archivematica/Issues#124