Fix incompatibility with pytest 3.7
The presence of `__init__.py` in the `storage_service` directory is causing
pytest to break with the following error:

    _pytest.nodes.Collector.CollectError: ImportError while importing test module
    Hint: make sure your test modules/packages have valid Python names.

Removing the file solved the problem. I think that pytest does not
expect `storage_service` to be a package. We don't treat it as a package
as far as I know.