Store AIP: fix pointer file creation if extension is unknown
Store AIP was raising when the value in `current_path` had an unknown
extension. The code responsible for creating the `premis:object` element
in the pointer file was assuming that the extension was either `.7z` or
`.bz2`. However when the space is using DSpace, the path does not include
the extension. This commit changes the lookup so `format_name` and
`format_registry_key` are only populated when the extension is known.

Connects to https://github.com/archivematica/Issues/issues/69.