Enable ptr file creation for remote filesystem
When a Storage Service is configured to connect to a pipeline on a
remote server it cannot see the location of the package before it
is transferred over to the storage service. In its current
configuration Archivematica tries to create a pointer file based on
an AIP being a file, or not, before this happens. We implement a
change to query the remote location and create a test for isfile().