Enable a region to be specified for a Bucket
The CreateBucketConfiguration is used to define the AWS region. The
bucket is created in the default AWS region when not using this
parameter.

We also test for a bucket's existence before attempting to create one.
We do so using get_bucket_location. If a 'NoSuchBucket' error is
raised then it prompts the creation of a new one.

Boto3 has been upgraded and pinned to the latest version.