Add gzip compression
This commit adds gzip compression support for pointer files as well
as adding missing 7z Copy compression support. Compression operations
are also refactored to provide a cleaner logic in package.py.

Tests have been created for much of this work as well.

Additionally, in the process of submitting this work, the storage
service was updated to support PREMIS3. As part of that work, and
this, we have made sure that PREMIS2 and PREMIS3 namespaces are
looked at when pointer files are used for reingest so that the
compression used can be identified as accurately as possible.