package: change size from Integer to BigInteger
Currently the /api/v2/files endpoint fails when trying to register files
that are bigger than 2.15 GB. This is due to the maximum value of the
size field in the Package model.

When the size field is configured as an IntegerField the maximum file
size of individual files in the AIP is limited to 2.15GB (2147483647
bytes). Changing this to BigInteger increases this limit to the order of
exabytes. Making the maximum allowed file size of the file system the
potential bottleneck.