Replicate AIP bag contents not container dir
Ensure that during replication we're copying the AIP (bag) contents
and not the container directory. Copying the container directory
pushes the replicated structure one level further down than we expect
and so when we come to test for something, e.g. fixity, we cannot
find the structural aspects of the bag, e.g. manifests, to perform
validation.