Add integration tests
This commit adds an additional level of testing that does not require us to set
up Archivematica pipelines. It is run on every commit and pull request made to
this repository using a GitHub workflow.

Presumably, this could be incorporated to our existing pytest suite using
custom markers. My suggestion is to decide that later on.

Some of the test scenarios will not pass until https://github.com/artefactual/archivematica-storage-service/pull/524 is merged. Connects to https://github.com/archivematica/Issues/issues/1149.