Add CAS authentication to Storage Service
This commits adds support for authentication via Central Authentication
Service (CAS) to the Storage Service. It includes configuration
options for auto-setting email addresses for users based on the rule
USERNAME@DOMAIN as well as for setting user.is_superuser based on the
presence or absence of configurable expected values in user attributes
returned by a CAS server during p3/serviceValidate.

Because the CAS middleware bypasses the Archivematica login screen and
thus prevents other single sign-on methods from being utilized, an
ImproperlyConfigured exception will be raised when attempting to start
Archivematica with CAS enabled in addition to Shibboleth or LDAP.