Correct delete and reingest issues w/ replication
A series of issues existed with replication in the storage service.
Packages would only be replicated correctly the first time they were
reingested and they were not updated when the package was updated,
e.g. through reingest.

More generally, deletion would not trickle down through the deletion
of a source package to the deletion of a source package's replicas.

The quad-directory folder structure we recognize in the AIPstore
would also not be deleted if a package was successfully deleted.

This commit collects a number of smaller commits that look to correct
each of these issues as well as add comprehensive testing around the
fixes.

Small pieces of refactoring have been completed along the way to
improve code-readability and the different interfaces that we are
working with.

Commit history for signaificant changes below:

-- Delete from local path tests

Provide some tests for the Storage Service's basic deletion
capabilities. The tests capture deletion of files and directories
as well as attempt to ensure the integrity of an AIPstore following
deletion. These tests should open the opportunity to increase the
complecity of deletion per archivematica/issues#359.

-- Remove quad-directory leafs from AIPstore

Previously we were not running any routine to clean up an AIPstore
when deleting an AIP. This code existed for pointer files. In this
commit we add that same capability for deleting AIPs. When an AIP
is deleted, the quad-directory structure it exists in will be deleted
up to the point another quad-directory is not empty, i.e. may contain
another AIP deeper down. Tests created in the previous commit are
extended to make sure the function behaves as expected.

-- Enable deletion of replicas with source package

When a source package is deleted from the storage service we will now
look for replicas and delete those at the same time, if the source
package is deleted successfully. Tests have been added to support
the use-cases and to enhance existing deletion tests.

Because we have added replicas to the fixtures some tests shifted.
The datatables tests have been updated to reflect the new changes
in a way that should be easier to change in future if this happens
again. The dynamic call by those tests to create replicas has been
replaced with the fixture data to test equivalent functionality.

The arkivum fixtures had packages whose pk enumerated from 10 and
11 which has the potential to clash with additions to our basic
packages.json. Those numbers now run backwards 99, 98, etc.

-- Ensure that replicas are created during reingest

Replicas are now recreated during reingest and testing has been added
from both a technical and digital preservation perspective to ensure
that the properties of new replicas are consistent with what we
expect.