Increase S3 timeout config and improve S3 logging
Boto3's default timeout was 60 seconds which was largely okay for
chunked data transfer across the wire from the storage service to an
S3 instance. Where we started to see more trouble with just a 60
second timeout was when the data sent between the two services is
increased, so for example, during the handshake at the end of an
extra large transfer. This commit sees the default timeout increased
to the value of one of the implementations we have tested against and
makes the timeout configurable from the environment. Timeouts
affected are read and connect timeouts used by Boto3 and S3.

Users should also be able to get more information out of the Boto3
library to help diagnose problems. They can use Archivematica's
defaults or override them. This commit preps the way for their, first,
knowing what is possible, and second, being able to do that easily.

Further, logging documentation is improved so that it isn't one foot
in the present and one in the past. It also adds notes about when
administrators or developers might be able to get more out of third
party libraries like S3. Finally we round off this commit with a call
to the community to help us to improve what is output by
Archivematica's logging through logging issues and PRs.

Additional changes have been made to the README to remove isolated
Backticks from configuration fields. Backticks were indicating short
blocks of code but with no content and this can be interpreted
strangely by the rendering application. We add single-quotes to
denote an empty string while maintaining the backtick markdown.