Add configurable S3 timeouts
Boto3's default timeout was 60 seconds which was largely okay for
chunked data transfer across the wire from the storage service to an
S3 instance. Where we started to see more trouble with just a 60
second timeout was when the data sent between the two services is
increased, so for example, during the handshake at the end of an
extra large transfer. This commit sees the default timeout increased
to the value of one of the implementations we have tested against and
makes the timeout configurable from the environment. Timeouts
affected are read and connect timeouts used by Boto3 and S3.