Avoid config reset in pipeline edit form
When a pipeline is being edited, "create_default_locations" defaults to
True. This is often misleading because it forces the pipeline to
re-create the default locations.

This commit changes the pipeline view so the checkbox appears enabled by
default when creating a new pipeline but not when editing an existing
one.

Connects to https://github.com/archivematica/Issues/issues/1107.