Start watchdog thread explicitly
Until now, the watchdog thread was started as a side effect of loading
the Django models at startup time. This was inconvenient because the
thread would run (and sometimes fail) when running migrations or other
management commands, running tests, etc...

This commit changes moves the code into a function that is explicitly
executed from `common.startup` when the WSGI application launched.

Connects to https://github.com/archivematica/Issues/issues/1433.