Store package checksum to verify correct package
This commit adds checksum and checksum_algorithm fields to the Package
model, which are populated during AIP storage. These are then used
during fixity checks to ensure that not only does the Bag validate, but
we are validating the expected Bag.

For compressed AIPs, this allows us to verify the fixity of the package
without needing to extract the AIP, resulting in much quicker fixity
checks.

For compressed packages, the checksum is the hash of the compressed
package file. For uncompressed packages, we store the hash of the Bag's
tagmanifest file. If multiple are present, we will use the first of:
sha512, sha256, or md5. This is intended to match the BagIt
specification's preference for sha512 and sha256, respectively.

This commit also includes a management command to populate the new
fields for existing AIPs that would otherwise have null values on
upgrade to Storage Service 0.20.x.