Fix gpg key generation in Rocky Linux
From https://docs.red-dove.com/python-gnupg/index.html#generating-keys:

If no passphrase is wanted for a key (which might be the default for
tests, say), or if you want to use an empty string as a passphrase,
then you should specify True for this parameter. Otherwise, and if you
don’t use pinentry to enter a passphrase, then GnuPG >= 2.1 will not
allow this.

This sets the `no_protection` argument when generating key input.