Generate SBOM in CI
Add ability to generate and upload an SBOM to Dependency Track in CI.

This CI workflow will autocreate a new Dependency Track project for any
qa or stable branch this workflow is picked to. The projectName
will be set to "archivematica" and the project version will be the branch
name.

Because the parentName is set to 'Archivematica', this SBOM will be grouped under
a project group of the same name which must be present in advance.

Dev dependencies will be included in SBOM.