Import helpers in local and production settings
The base settings module used to import the get_env_variable helper
but didn't use it. It was indirectly imported from the local and
production settings modules.

When Ruff was initially run on the repository it detected the unused
import and "fixed" it breaking the other modules.

This imports the helpers needed explicitly from each settings module
and it also moves the ignored "noqa" pragmas into Ruffs configuration
file.