Update get_7z_version helper for compatibility with newer 7z versions
For reference this is the top ouput of the 7z command in version
16.02:

```console

7-Zip [64] 16.02 : Copyright (c) 1999-2016 Igor Pavlov : 2016-05-21
p7zip Version 16.02 (locale=C,Utf16=off,HugeFiles=on,64 bits,4 CPUs Intel(R) Core(TM) i7-7700HQ CPU @ 2.80GHz (906E9),ASM,AES-NI)
```

We currently take only the second line.

And this is the output of the 7z command in version 23.01:

```console

7-Zip 23.01 (x64) : Copyright (c) 1999-2023 Igor Pavlov : 2023-06-20
 64-bit locale=C.UTF-8 Threads:4 OPEN_MAX:1048576
```

This commit updates the helper function to keep the old behavior and
to get and merge both lines in version 23.01 and later.