Check rclone container existence more efficiently
In the rclone model's _ensure_container_exists method, rclone ls is used
to check whether a container exists on an rclone remote. If not, rclone
mkdir is used to create it. rclone ls is recursive by default, however,
so this call can take minutes if the container is large.
This commit replaces the rclone ls call with rclone lsf --dirs-only,
which lists only root level directories and therefore runs much
faster.